requirements.txt
anthropic>=0.40.0
requests>=2.31.0
matplotlib>=3.8.0
pyyaml>=6.0
python-dotenv>=1.0.0
